remove the Lightspeed LSP server
