+++ packages/ansible-language-server/package.json
@@ -48,6 +48,7 @@
   "dependencies": {
     "@flatten-js/interval-tree": "^1.1.2",
     "antsibull-docs": "^1.0.0",
+    "axios": "^1.6.7",
     "glob": ">=9.3.2",
     "ini": "^4.1.1",
     "lodash": "^4.17.21",
