+++ src/extension.ts
@@ -73,7 +73,6 @@ import { LightspeedAuthSession } from "./interfaces/lightspeed";
 import { showPlaybookGenerationPage } from "./features/playbookGeneration/playbookGenerationPage";
 
 export let client: LanguageClient;
-export let lsClient: LanguageClient;
 export let lightSpeedManager: LightSpeedManager;
 export const globalFileSystemWatcher: IFileSystemWatchers = {};
 
@@ -158,7 +157,6 @@ export async function activate(context: ExtensionContext): Promise<void> {
   lightSpeedManager = new LightSpeedManager(
     context,
     client,
-    lsClient,
     extSettings,
     telemetry,
   );
@@ -250,7 +248,7 @@ export async function activate(context: ExtensionContext): Promise<void> {
       async () => {
         await playbookExplanation(
           context.extensionUri,
-          lsClient,
+          client,
           lightSpeedManager.lightSpeedAuthenticationProvider
         );
       }
@@ -546,36 +544,6 @@ const startClient = async (
     clientOptions,
   );
 
-  const lsServerModule = context.asAbsolutePath(
-    path.join("out", "client", "src", "features", "lightspeed", "server.js")
-  );
-
-  const lsServerOptions: ServerOptions = {
-    run: { module: serverModule, transport: TransportKind.ipc },
-    debug: {
-      module: lsServerModule,
-      transport: TransportKind.ipc,
-    },
-  };
-
-  const lsClientOptions: LanguageClientOptions = {
-    // register the server for Ansible documents
-    documentSelector: [{ scheme: "file", language: "ansible" }],
-    revealOutputChannelOn: RevealOutputChannelOn.Never,
-    errorHandler: telemetryErrorHandler,
-    outputChannel: new TelemetryOutputChannel(
-      outputChannel,
-      telemetry.telemetryService
-    ),
-  };
-
-  lsClient = new LanguageClient(
-    "lightSpeedServer",
-    "Ansible LightSpeed Server",
-    lsServerOptions,
-    lsClientOptions
-  );
-
   context.subscriptions.push(
     client.onTelemetry((e) => {
       telemetry.telemetryService.send(e);
@@ -584,7 +552,6 @@ const startClient = async (
 
   try {
     await client.start();
-    await lsClient.start();
 
     // If the extensions change, fire this notification again to pick up on any association changes
     extensions.onDidChange(() => {
